pyproject.toml
[tool.poetry]
name = "spartan-framework"
version = "0.1.8"
description = "A micro framework for serverless development in AWS"
authors = ["Sydel Palinlin <sydel.palinlin@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
annotated-types = ">=0.6.0"
colorama = "*"
cryptography = ">=3.3.1"
greenlet = "!=0.4.17"
httpx = "0.28.1"
iniconfig = "*"
jmespath = ">=1.0.1,<2.0.0"
pluggy = ">=1.5,<2"
pydantic-core = "^2.37.2"
pydantic-settings = "2.10.1"
pydantic = "^2.11.7"
python-dateutil = ">=2.1,<3.0.0"
python-dotenv = ">=0.21.0"
python-jose = "3.5.0"
pytz = "2025.2"
requests = ">=2.5"
responses = ">=0.13.0"
sqlalchemy = "2.0.43"
typing-extensions = ">=4.6.0"
werkzeug = ">=0.5,<4.0.0"
xmltodict = "*"
pymysql = "1.1.2"
pg8000 = "1.31.4"
psycopg2-binary = "2.9.10"
email-validator = "^2.0.0"
jsonpickle = "^4.1.1"



[tool.poetry.group.dev.dependencies]
alembic = "1.16.5"
Mako = "*"
packaging = "*"
poetry-plugin-export = "^1.9.0"
rich = "^14.1.0"
autoflake = "^2.2.1"
aws-lambda-powertools = "^3.19.0"
aws-xray-sdk = "^2.12.0"
boto3 = "^1.40.19"
botocore = "^1.40.19"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
bandit = "^1.7.5"
pytest-mock = "^3.14.1"
pytest-asyncio = "^1.1.0"
faker = "^37.6.0"
mocker = "^1.1.1"
moto = "^5.0.0"
tox = "^4.11.0"
mypy = "^1.5.0"
types-requests = "^2.31.0"
types-python-dateutil = "^2.8.0"
typer = "^0.16.0"
pytest-xdist = "^3.8.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.serverless
  | \.vscode
)/
'''

[tool.autopep8]
max-line-length = 120
exclude = [
    "app/exceptions/__init__.py",
    ".git",
    "build",
    "dist",
    ".venv",
    ".serverless",
    "node_modules",
    ".vscode",
    ".pytest_cache",
    "__pycache__"
]

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
recursive = true
in-place = true
exclude = [
    "app/exceptions/__init__.py",
    ".git",
    "build",
    "dist",
    ".venv",
    ".serverless",
    "node_modules",
    ".vscode",
    ".pytest_cache",
    "__pycache__"
]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
lines_after_imports = 2
skip = [
    "app/exceptions/__init__.py",
    ".git",
    "build",
    "dist",
    ".venv",
    ".serverless",
    "node_modules",
    ".vscode",
    ".pytest_cache",
    "__pycache__",
    ".tox",
    ".vscode",
    "htmlcov",
    "coverage.xml",
    ".mypy_cache",
    ".pytest_cache",
]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
    "integration",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
